fix clang-tidy 2

+++ tasks/badanov_a_select_edge_sobel/tests/functional/main.cpp
@@ -224,7 +224,7 @@ TEST_F(BadanovASelectEdgeSobelGradientTests, EdgeDetectionWorksWithSharpTransiti
 
   auto output = task.GetOutput();
   ASSERT_FALSE(output.empty());
-  ASSERT_GT(output.size(), 4);
+  ASSERT_GT(output.size(), static_cast<size_t>(4));
   EXPECT_GT(static_cast<int>(output[4]), 0);
 }
